Ensure ADD statements always have same order in generaated Dockerfile
The list of included files returned from DirectoryScanner can be in a different order
each time. This causes the ADD statements in the generated Dockerfile to appear in a
different order. We want to avoid this so each run of the plugin always generates the same
Dockerfile, which also makes testing easier. Sort the list of paths for each resource
before adding it to the allCopiedPaths list. This way we follow the ordering of the
resources in the pom, while making sure all the paths of each resource are always in the
same order.