Write digest to image info file on push
This adds a new filed to the image info file named "digest", which
contains the digest of the image if it was returned by the registry
to which it was pushed. The digest field does NOT include the image
name. I.e. it's "sha256:12ae23" rather than "name@sha256:12ae23".