correct the base path of files in BuildMojo tests
the test pom.xml's used in BuildMojo refer to `src/test/resources/FILE`
but since the Maven project used with the Mojo has a base directory of
`src/test/resources` already, this ends up with this portion of the path
appearing twice in the final result.

This was masked by the homegrown file copying logic before but now
surfaces when using MavenResourcesFiltering.