ensure entire directories are copied for ADDs
Prior to this change, a mojo configuration like

```
<resource>
  <directory>some/directory</directory>
  <targetPath>/data</targetPath>
</resource>
```

would generate an `ADD /data /data` instruction but the `some/directory`
directory is never copied into the Docker build context directory, so
the actual build will fail.

With this change, the directory is copied to the build context
directory.

Note: it might seem weird that the `targetPath` is used as the directory
name to use *inside* the build context as opposed to just the target
path inside the container, but this is to mirror behavior when copying
individual files.

Fixes #146.