update maven libraries and remove guava restriction
Remove the pinning of Guava to version 14.0.1 to fix
https://github.com/spotify/docker-maven-plugin/issues/165.

Guava was held to 14.0.1 before because a combination of Plexus/Guice
code used or invoked by the maven plugin testing harness called methods
in Guava that have been removed after version (specifically
`MapMaker.makeComputingMap).

The main fix to remove that need is to avoid using that part of the
testing harness altogether by stubbing out the `MavenProject` instance
instead of trying to look it up in the Plexus container.

Fixes #165.